requirements.txt
# Core Dependencies
pyyaml>=6.0.1
python-dotenv>=1.0.0
fastapi[all]
# Email Processing
email-validator>=2.1.0
icalendar>=5.0.11

# Text Processing & Extraction
beautifulsoup4>=4.12.2
lxml>=4.9.3
html5lib>=1.1
spacy>=3.7.2
phonenumbers>=8.13.27

# GLiNER - Modern NER (requires transformers and torch)
gliner>=0.2.0
transformers>=4.35.0
torch>=2.0.0
tokenizers>=0.15.0
huggingface-hub>=0.19.0

# Machine Learning (for filters)
scikit-learn>=1.3.2
joblib>=1.3.2
numpy>=1.24.0
pandas>=2.0.0

# Database
mysql-connector-python>=8.2.0

# Utilities
requests>=2.31.0
tqdm>=4.66.1
python-dateutil>=2.8.2
regex>=2023.10.3
tldextract>=3.4.0
